162. Dropdown Architecture 163. Scaffolding the Dropdown(Section 12: Understanding Hooks in React)
We are now all done with our little search widget, Sorina, start to move forward to our next one,

which is going to be a little dropdown.

And the overall goal of this dropdown is to get some more practice with you, state use effect and a

new hook that we're going to discuss called use RF.

The overall goal of the dropdown itself is to show some list of options the user is going to click on.

This little menu is going to open up.

And for this initial example, we're just gonna throw in some different colors that a user can select.

The red, green and blue.

Whenever a user changes a color selection, we're then going to update some text color immediately after

it.

That's pretty much it.

Now, the dropdown itself, we are going to make reusable, unlike the search over here.

So this little search widget, it was technically reusable, but it didn't really accept or receive

any props.

And so there was a limited ability for us to somehow customize how it worked with the dropdown.

On the other hand, we definitely got to make this thing reusable.

And that means that we need to be able to customize the options that are provided inside the dropdown.

So let's take a look at an architecture diagram or two and understand how we're gonna put this thing

together.

All right.

So we're once again going to have our app component.

It's going to show this dropdown, the app component is going to have a list of options.

So these are going to be the different options a user can select in the drop down menu.

These options will be provided as props down to the drop down component.

And the drop dropdown is going to use that set of options to decide what to show.

The options itself is going to be an array of objects.

So this might be a what the data structure looks like, there'll be an array of objects where every

object has a label and a value.

So in this case, if we pass down an array of objects that have red, green and blue, whenever a user

opens up that dropdown, we should show options.

Red, green and blue.

You might be curious why we have both a label and a value property.

Well, it's really straightforward when we show some text to the user right here.

We might want to not precisely say something like red.

We might want to say something like the color red and say the color green.

And a shade of blue.

The label is what actually gets displayed to the user.

The value is how we interpret whatever the user selects.

So a value of simply red is a lot easier for you and I to use in some way, then try to pass this label

and understand what the user just selected.

That's why we have both a label and a value.

Besides this set of props, we are also gonna have some state aid to this component as well.

We're going to call the states or the single piece of state selection and selection is going to record

what they currently selected.

Option is selection is going to exist inside of the parent component, which in this case is app.

And we're going to provide that down as a prop as well to the dropdown to tell it what it should currently

have selected.

You should really think of this as kind of like the relationship between a component and an input field

where, remember, we've got some kind of value that we specify on the input to tell it what the current

value of the input is.

Very similar for the dropdown here as well.

All right.

Let's just kind of a quick overview on some of the things we need to be aware of.

Now, if any of the stuff right now looks unclear, particularly around this selection, bit fully fine.

As soon as we start to write out some code, you'll have a good idea of what's going on.

Let's pause right here and then start to put together this dropdown in just a moment.

163. Scaffolding the Dropdown

Back inside of our code editor, let's create a new file inside of our components directory called DropDown

Dot J.S..

As usual, we'll put together just a little bit of boilerplate inside of here and then show that component

inside of our app component.

So at the very top, I will import react from react.

Well, then make a dropdown.

And right now, as usual, the simple H1.

It says dropdown.

And an export default dropped down at the bottom.

Let's now go back over to our APT.

James file.

Here's APJ, yes.

At the very top, we will import.

Once we've got that important place, well, then go down to the component itself and once again right

now, we're just going to manually change which component you're are displaying again later on as we

finish up all these different widgets.

We will eventually introduce some navigation between them.

So right now, we'll just delete, search and replace it with dropdown.

Let's save that and just do a quick refresh of our app over here.

And make sure that we see dropped down.

Very good.

All right.

That's simple enough.

Now, the next thing we're going to do is put together that options array.

No options, as we just discussed, is going to be an array of objects.

We're going to define options inside of our app component and then pass it down to the dropdown as props.

And it's going to specify what different options we want to show inside the dropdown.

So back inside of our app component file.

I'm going to declare right above the component self.

An array of options.

Now we can define this inside the component, outside the component.

It really doesn't make a difference because this is going to be a static array of data.

It's not going to change in any way.

So it doesn't matter if we define this inside the component or outside.

Inside of options, we'll put together three different objects and each one will have a label and a

value.

We will first put in a label of the color red.

The value of red.

A label of the color green.

With a value of green.

And then finally.

Label of a shade of blue, but the value of.

Now we can take this array of options and pass it down as a property into our drop down component.

We'll say options is options.

The whole goal behind this is to make sure that we can make use of the drop down component in different

places inside of application and have them show a different list of options to the user.

That's why we are defined this as props are we're gonna pass down.

So it looks good.

Let's take a pause here.

When I come back, the next video, we're going to start to add in a lot more jail sex inside of our

drop down component self.

We're going to map over that list of options and show it inside of some nice looking menu.